pyproject.toml
[project]
name = "calculator-mcp"
version = "0.1.0"
description = "MCP tool for calculator"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=0.5.0",
    "mcp>=1.9.0",
    "python-dotenv>=1.1.0",
    "starlette>=0.46.2",
    "uvicorn>=0.34.2",
]
